fix: 有无符号数之间的显式转换

+++ brightness_manager/src/brightness_dimming.cpp
@@ -88,8 +88,8 @@ void BrightnessDimming::StartDimming(uint32_t from, uint32_t to, uint32_t durati
     }
     DISPLAY_HILOGI(FEAT_BRIGHTNESS, "animation orig mTotalSteps=%{public}d", static_cast<int32_t>(mTotalSteps));
     if (mTotalSteps > 1 && abs(mStride) >= 1) {
-        mTotalSteps = static_cast<uint32_t>((abs(changeBrightness) - abs(mStride) *
-            mTotalSteps)) / abs(mStride) + mTotalSteps;
+        mTotalSteps = static_cast<uint32_t>(static_cast<uint32_t>(abs(changeBrightness)) -
+            static_cast<uint32_t>(abs(mStride)) * mTotalSteps) / static_cast<uint32_t>(abs(mStride)) + mTotalSteps;
         DISPLAY_HILOGI(FEAT_BRIGHTNESS, "animation update mTotalSteps=%{public}d", static_cast<int32_t>(mTotalSteps));
     }
     mCurrentStep = 0;
